+++ src/components/ui/govbr-switch/index.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import Box from '@mui/material/Box'
 import FormControlLabel from '@mui/material/FormControlLabel'
 import Switch from '@mui/material/Switch'